# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Beacon Nonprofit Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /safehouses:
    get:
      operationId: listSafehouses
      tags: [safehouses]
      summary: List all safehouses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of safehouses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafehouseListResponse"
    post:
      operationId: createSafehouse
      tags: [safehouses]
      summary: Create safehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSafehouseBody"
      responses:
        "201":
          description: Created safehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Safehouse"
  /safehouses/{id}:
    get:
      operationId: getSafehouse
      tags: [safehouses]
      summary: Get safehouse by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Safehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Safehouse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSafehouse
      tags: [safehouses]
      summary: Update safehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSafehouseBody"
      responses:
        "200":
          description: Updated safehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Safehouse"
    delete:
      operationId: deleteSafehouse
      tags: [safehouses]
      summary: Delete safehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List residents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: safehouseId
          in: query
          schema:
            type: integer
        - name: caseStatus
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
        - name: assignedWorkerId
          in: query
          schema:
            type: integer
        - name: reintegrationStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resident list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentListResponse"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "201":
          description: Created resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get resident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResidentBody"
      responses:
        "200":
          description: Updated resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete resident record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /residents/stats:
    get:
      operationId: getResidentStats
      tags: [residents]
      summary: Aggregate resident statistics
      parameters:
        - name: safehouseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Resident stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResidentStats"
  /residents/{id}/timeline:
    get:
      operationId: getResidentTimeline
      tags: [residents]
      summary: Get full timeline of case events for a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"

  # ─── PROCESS RECORDINGS ───────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]

    # ─── AUTH ───────────────────────────────────────────────────────────────
...
    Safehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        currentOccupancy:
          type: integer
        programAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, location, capacity, currentOccupancy, status]
    CreateSafehouseBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        programAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        contactName:
          type: string
        contactEmail:
          type: string
      required: [name, location, capacity]
    UpdateSafehouseBody:
      type: object
      properties:
        name:
          type: string
        location:
          type: string
        capacity:
          type: integer
        programAreas:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive]
        contactName:
          type: string
        contactEmail:
          type: string
    SafehouseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Safehouse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
...
    Resident:
      type: object
      properties:
        id:
          type: integer
        residentCode:
          type: string
        safehouseId:
          type: integer
        safehouseName:
          type: string
        assignedWorkerId:
          type: ["integer", "null"]
        assignedWorkerName:
          type: ["string", "null"]
        caseStatus:
          type: string
          enum: [active, closed, transferred, deceased]
        caseCategory:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        reintegrationStatus:
          type: string
          enum: [not_started, in_progress, ready, completed]
        admissionDate:
          type: string
        dischargeDate:
          type: ["string", "null"]
        ageGroup:
          type: ["string", "null"]
        lastUpdated:
          type: string
        createdAt:
          type: string
      required: [id, residentCode, safehouseId, caseStatus, caseCategory, riskLevel, reintegrationStatus, admissionDate]
    CreateResidentBody:
      type: object
      properties:
        residentCode:
          type: string
        safehouseId:
          type: integer
        assignedWorkerId:
          type: integer
        caseStatus:
          type: string
          enum: [active, closed, transferred, deceased]
        caseCategory:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        reintegrationStatus:
          type: string
          enum: [not_started, in_progress, ready, completed]
        admissionDate:
          type: string
        ageGroup:
          type: string
      required: [residentCode, safehouseId, caseStatus, caseCategory, riskLevel, reintegrationStatus, admissionDate]
    UpdateResidentBody:
      type: object
      properties:
        safehouseId:
          type: integer
        assignedWorkerId:
          type: integer
        caseStatus:
          type: string
        caseCategory:
          type: string
        riskLevel:
          type: string
        reintegrationStatus:
          type: string
        dischargeDate:
          type: string
        ageGroup:
          type: string
    ResidentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Resident"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    ResidentStats:
      type: object
      properties:
        totalActive:
          type: integer
        newAdmissions:
          type: integer
        casesNeedingUpdate:
          type: integer
        highRiskResidents:
          type: integer
        riskDistribution:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
              count:
                type: integer
            required: [level, count]
        statusDistribution:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              count:
                type: integer
            required: [status, count]
      required: [totalActive, newAdmissions, casesNeedingUpdate, highRiskResidents]
    TimelineEvent:
      type: object
      properties:
        id:
          type: string
        eventType:
          type: string
        eventDate:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        actor:
          type: ["string", "null"]
        severity:
          type: ["string", "null"]
      required: [id, eventType, eventDate, title]

    # ─── PROCESS RECORDING ────────────────────────────────────────────────────
...